Fix API connection issue
Fixed an issue where the front-end was using the wrong API_URL.
The issue is because when Vue apps are built, env vars that don't
start with 'VUE_APP_' are ignored by the compiler. To fix the
issue, I changed API_URL to VUE_APP_API_URL in .env.* files and
corrected references to the env var.